Converter for SIPM of HiDRA taking shape
Work is ongoing. I have introduced a CMakeLists.txt for the SIPM package, and I have started to add a number of classes and functionalities.
For the moment I am focusing only on teh Spectroscopy and Timing mode. but I hope the code is flexible enough to be extended without too much work.